Finally, move plugins in the correct folder
