+++ RecoLocalCalo/HGCalRecProducers/plugins/HGCalImagingAlgo.cc
@@ -1,4 +1,4 @@
-#include "RecoLocalCalo/HGCalRecAlgos/interface/HGCalImagingAlgo.h"
+#include "RecoLocalCalo/HGCalRecProducers/interface/HGCalImagingAlgo.h"
 
 // Geometry
 #include "DataFormats/HcalDetId/interface/HcalSubdetector.h"
