+++ RecoLocalCalo/HGCalRecProducers/plugins/HGCalLayerClusterPluginFactory.cc
@@ -1,7 +1,7 @@
-#include "RecoLocalCalo/HGCalRecAlgos/interface/HGCalLayerClusterAlgoFactory.h"
-#include "RecoLocalCalo/HGCalRecAlgos/interface/HGCalClusteringAlgoBase.h"
-#include "RecoLocalCalo/HGCalRecAlgos/interface/HGCalImagingAlgo.h"
-#include "RecoLocalCalo/HGCalRecAlgos/interface/HGCalCLUEAlgo.h"
+#include "RecoLocalCalo/HGCalRecProducers/interface/HGCalLayerClusterAlgoFactory.h"
+#include "RecoLocalCalo/HGCalRecProducers/interface/HGCalClusteringAlgoBase.h"
+#include "RecoLocalCalo/HGCalRecProducers/interface/HGCalImagingAlgo.h"
+#include "RecoLocalCalo/HGCalRecProducers/interface/HGCalCLUEAlgo.h"
 #include "FWCore/ParameterSet/interface/ValidatedPluginFactoryMacros.h"
 #include "FWCore/ParameterSet/interface/ValidatedPluginMacros.h"
