+++ RecoLocalCalo/HGCalRecProducers/python/hgcalLayerClusters_cff.py
@@ -10,10 +10,10 @@ from SimCalorimetry.HGCalSimProducers.hgcalDigitizer_cfi import fC_per_ele, hgce
 
 hgcalLayerClusters = hgcalLayerClusters_.clone()
 
-hgcalLayerClusters.dEdXweights = cms.vdouble(dEdX.weights)
-hgcalLayerClusters.fcPerMip = cms.vdouble(HGCalUncalibRecHit.HGCEEConfig.fCPerMIP)
-hgcalLayerClusters.thicknessCorrection = cms.vdouble(HGCalRecHit.thicknessCorrection)
-hgcalLayerClusters.fcPerEle = cms.double(fC_per_ele)
 hgcalLayerClusters.timeOffset = hgceeDigitizer.tofDelay
-hgcalLayerClusters.noises = cms.PSet(refToPSet_ = cms.string('HGCAL_noises'))
-hgcalLayerClusters.noiseMip = hgchebackDigitizer.digiCfg.noise_MIP
+hgcalLayerClusters.plugin.dEdXweights = cms.vdouble(dEdX.weights)
+hgcalLayerClusters.plugin.fcPerMip = cms.vdouble(HGCalUncalibRecHit.HGCEEConfig.fCPerMIP)
+hgcalLayerClusters.plugin.thicknessCorrection = cms.vdouble(HGCalRecHit.thicknessCorrection)
+hgcalLayerClusters.plugin.fcPerEle = cms.double(fC_per_ele)
+hgcalLayerClusters.plugin.noises = cms.PSet(refToPSet_ = cms.string('HGCAL_noises'))
+hgcalLayerClusters.plugin.noiseMip = hgchebackDigitizer.digiCfg.noise_MIP
